requirements.txt
asgiref==3.7.2
Django==4.2.5
django-localflavor==4.0
djangorestframework==3.14.0
python-dotenv==1.0.0
python-stdnum==1.19
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
